# [PERF] mail: speed up fetching of `avatar_128`